// connectors/slack/types.ts
export interface WorkspaceRow {
	DATA: Uint8Array | string | null;
	ENTERPRISE_ID: string | null;
	ID: number;
	TEAM: string | null;
	TEAM_ID: string | null;
	URL: string | null;
	USER_ID: string | null;
	USERNAME: string | null;
}

export interface ChannelRow {
	data: Uint8Array | string | null;
	id: string;
	name: string | null;
}

export interface ChannelUserRow {
	CHANNEL_ID: string;
	USER_ID: string;
}

export interface UserRow {
	data: Uint8Array | string | null;
	id: string;
	username: string | null;
}

export interface MessageRow {
	CHANNEL_ID: string;
	DATA: Uint8Array | string | null;
	IS_PARENT: number | null;
	NUM_FILES: number | null;
	THREAD_TS: string | null;
	TS: string;
	TXT: string | null;
}